middleware/kubernetes: pull TXT out of parseRequest
Put the TXT handling one layer higher and remove it from parseRequest.
Also rename the podsvc field in there to podOrSvc. Now that it isn't
used anymore for TXT record (dns-version) was put in there. We can make
this a boolean.

Started new parseRequest called parse and add text for it. Removed
superfluous tests.