Make k.Namespace a map
This remove pkg/strings as it is not used anymore. Also add a few
helper function to make unexposed namespaces easier to see in the
code.unexposed namespaces easier to see in the code.unexposed namespaces
easier to see in the code.unexposed namespaces easier to see in the
code.