mw/health: poll other middleware
This add the infrastructure to let other middleware report their health
status back to the health middleware. A health.Healther interface is
introduced and a middleware needs to implement that. A middleware
that supports healthchecks is statically configured, but this could be
made into a RegisterHealth thing so it can *also* be done by the
middleware implementing it.

Every second each supported middleware is queried and the global health
state is updated.

The erratic middleware implements Healther as an example and updates its
documentation to say so.