plugin/{file,auto}: improve reload robustness
The reload (fsnotify) mechanism seems to "stop working" for some users,
it has been hard to pin down, what exactly. This PR makes a couple of
changes to help (hopefully) with this:

* Add more logging, including a panic (on startup) when the fsnotify
  can't be set
* Makes the fsnotify.Watcher global (we only need one), and keeps track
  a what directories we watch, as there may be a limit (there is a limit
  somewhere) to how many watcher you can have per directly.
* Only tries reloading a zone when there was a *write* to the zone, so
  moves are not seen anymore.

Testing this was a lot of zones is tricky-ish, so current test-suite
hasn't been updated. This might happen in subsequent comments because
the `watcher` struct does need unit tests.