healthcheck: move to DNS checks
This removes HTTP healthchecks and moves to DNS (via tcp) system.
This directly queries the upstream DNS server by using a . NS
no-recursive query. Any non-error DNS rcode ?? TODO(miek): if this makes
sense is taken as success.

The proxy and lookup code does not set an upstream unhealthy anymore,
but relies on the healthcheck to do is things.

Obviously tests fails.