Remove dnsutil.Dedup
Remove the code and remove the call in etcd and kubernetes handlers.
This does mean we should not add dups in the first place, which means
adding another map in backend_lookup to prevent dups from begin added.

This should cut doesn on the allocation because dnsutil.Dedup is very
expensive by converting everything to strings, we avoid doing that now.

Maybe/probably should add the same debug checks in backend_lookup.A and
backend_lookup.AAAA as well.