Default to scrubbing replies in the server
Every plugin needs to deal with EDNS0 and should call Scrub. Move this
into the server and wrapping the ResponseWriter into a ScrubWriter that
handles these bits for us. Result:

Less code and faster, because multiple chained plugins could all be
calling scrub and SizeAndDo - now there is just one place.

Most tests in file/* and dnssec/* needed adjusting because in those unit
tests you don't see OPT RRs anymore. The DNSSEC signer was also looking
at the returned OPT RR to see if it needed to sign - as those are now
added by the server (and thus later), this needed to change slightly.

Fixes: #2010

Signed-off-by: Miek Gieben <miek@miek.nl>