server: cancel context after 501 milliseconds
The context in each request is canceled after 501 ms. Plugins that check
ctx.Done() can abort (return) from whatever they are executing because
the waiting client will never see the reply.

The irony of this approach is that the cancelation is done by starting
another go-routine. It would be cooler if we would only start that
routine after x seconds, but doing that requires a go routine.

I don't believe Go provides out-of-the-box feature to look at the
packet's timestamp. Even if it did it requires each plugin to look at
it, the context.WithTimeout is somewhat cleaner.

Signed-off-by: Miek Gieben <miek@miek.nl>